+++ Source/ACE/Entity/BroadcastAction.cs
@@ -12,6 +12,8 @@
 
         PlayParticleEffect,
 
-        MovementEvent
+        MovementEvent,
+
+        BroadcastDeath
     }
 }