+++ Source/ACE/Entity/Landblock.cs
@@ -19,6 +19,8 @@ using ACE.Network.Enum;
 using ACE.Entity.Enum.Properties;
 using ACE.Network.Sequence;
 using ACE.Factories;
+using ACE.Network.GameEvent;
+using System.Diagnostics;
 
 namespace ACE.Entity
 {
@@ -342,6 +344,12 @@ namespace ACE.Entity
             Broadcast(args, true, Quadrant.All);
         }
 
+        public void HandleDeathMessage(WorldObject sender, DeathMessageArgs deathMessageArgs)
+        {
+            BroadcastEventArgs args = BroadcastEventArgs.CreateDeathMessage(sender, deathMessageArgs);
+            Broadcast(args, false, Quadrant.All);
+        }
+
         /// <summary>
         /// handles broadcasting an event to the players in this landblock and to the proper adjacencies
         /// </summary>
@@ -401,6 +409,14 @@ namespace ACE.Entity
                         // supresss updating if player is out of range of world object.= being updated or created
                         players = GetWorldObjectsInRange(wo, maxobjectRange, true).OfType<Player>().ToList();
                         Parallel.ForEach(players, p => p.SendMovementEvent(args.Motion, args.Sender));
+                        break;
+                    }
+                case BroadcastAction.BroadcastDeath:
+                    {
+                        // players never need an update of themselves
+                        // TODO: Filter to players in range and include adjacencies
+                        players = players.Where(p => p.Guid != args.Sender.Guid).ToList();
+                        Parallel.ForEach(players, p => p.BroadcastPlayerDeath(args.DeathMessage.Message, args.DeathMessage.Victim, args.DeathMessage.Killer));
                         break;
                     }
             }
@@ -539,6 +555,29 @@ namespace ACE.Entity
         {
             switch (action.ActionType)
             {
+                case GameActionType.TalkDirect:
+                    {
+                        // TODO: remove this hack (using TalkDirect) ASAP
+                        var g = new ObjectGuid(action.ObjectId);
+                        WorldObject obj = (WorldObject)player;
+                        if (worldObjects.ContainsKey(g))
+                        {
+                            obj = worldObjects[g];
+                        }
+                        DeathMessageArgs d = new DeathMessageArgs(action.ActionBroadcastMessage, new ObjectGuid(action.ObjectId), new ObjectGuid(action.SecondaryObjectId));
+                        HandleDeathMessage(obj, d);
+                        break;
+                    }
+                case GameActionType.TeleToHouse:
+                case GameActionType.TeleToLifestone:
+                case GameActionType.TeleToMansion:
+                case GameActionType.TeleToMarketPlace:
+                case GameActionType.TeleToPkArena:
+                case GameActionType.TeleToPklArena:
+                    {
+                        player.Teleport(action.ActionLocation);
+                        break;
+                    }
                 case GameActionType.ApplyVisualEffect:
                     {
                         var g = new ObjectGuid(action.ObjectId);