+++ Source/ACE/Entity/Portal.cs
@@ -15,9 +15,36 @@ namespace ACE.Entity
         {
         }
 
-        public Portal(AceObject aceObject)
-            : base(aceObject)
+        public Portal(AceObject aceO)
+            : base((ObjectType)aceO.TypeId, new ObjectGuid(aceO.AceObjectId))
         {
+            this.Name = aceO.Name;
+            this.DescriptionFlags = (ObjectDescriptionFlag)aceO.WdescBitField;
+            this.Location = aceO.Position;
+            this.WeenieClassid = aceO.WeenieClassId;
+            this.WeenieFlags = (WeenieHeaderFlag)aceO.WeenieFlags;
+
+            this.PhysicsData.MTableResourceId = aceO.MotionTableId;
+            this.PhysicsData.Stable = aceO.SoundTableId;
+            this.PhysicsData.CSetup = aceO.ModelTableId;
+
+            // this should probably be determined based on the presence of data.
+            this.PhysicsData.PhysicsDescriptionFlag = (PhysicsDescriptionFlag)aceO.PhysicsBitField;
+            this.PhysicsData.PhysicsState = (PhysicsState)aceO.PhysicsState;
+
+            this.PhysicsData.ObjScale = aceO.ObjectScale;
+
+            // game data min required flags;
+            this.Icon = (ushort)aceO.IconId;
+
+            this.GameData.Usable = (Usable)aceO.Usability;
+            this.GameData.RadarColour = (RadarColor)aceO.BlipColor;
+            this.GameData.RadarBehavior = (RadarBehavior)aceO.Radar;
+            this.GameData.UseRadius = aceO.UseRadius;
+
+            aceO.AnimationOverrides.ForEach(ao => this.ModelData.AddModel(ao.Index, (ushort)ao.AnimationId));
+            aceO.TextureOverrides.ForEach(to => this.ModelData.AddTexture(to.Index, (ushort)to.OldId, (ushort)to.NewId));
+            aceO.PaletteOverrides.ForEach(po => this.ModelData.AddPalette(po.SubPaletteId, po.Offset, po.Length));
         }
 
         public override void OnCollide(Player player)