+++ changelog.md
@@ -2,6 +2,14 @@
 [StackOverflow]
 * Streaming World Objects - Object now stay in sync for tracking.
 * they remove when out of range, readd when come back into range
+# ACEmulator Change Log
+### 2017-04-21
+[fantoms]
+* Removed `@reset-pos` after it was incorrectly re-added in another members commit.
+
+### 2017-04-21
+[Ripley]
+* Kill `Session.Network.EnqueueSend(new GameMessageUpdateObject(worldObject));` until we get proper movement implemented.
 
 ### 2017-04-20
 [Lidefeath]
@@ -12,10 +20,14 @@
 * Fixed PackedDWORD logic
 
 [Ripley]
+* Filled out Portal object and included ObjScale.
 * Changed ModelData.Serialize to use WritePackedDwordOfKnownType for PaletteGuid, palette.PaletteId, texture.OldTexture, texture.NewTexture and model.ModelID.
 * Changed IconOverlay and IconUnderlay to use WritePackedDwordOfKnownType.
 
 ### 2017-04-19
+[Mogwai-AC]
+* Added PackedDWORD
+
 [Ripley]
 * Changed WeenieClassid and Icon to use PackedDWORD in WorldObject.SerializeCreateObject
 * Updated PhysicsData to create and send a new currentMotionState when encountering a null one when flag PhysicsDescriptionFlag.Movement is set.
@@ -61,6 +73,18 @@
 * Added description and usage information to commands currently implemented.
 * Added sending chat messages to inform players upon logging in-game about the existence of ACE specific help.
 * Updated GameActionTalk to return better information when a command is invalid or is missing parameters.
+
+[fantoms]
+* Added `Queued` Teleporting.
+* Added landblock death message broadcasting.
+* Added lame static player death messages.
+* Now saves last death position in character positions.
+* Added Psuedo Event Action for internal ACE `GameActionEvent = 0xF819`.
+* Updated `GameEventYourDeath`.
+* Added `PlayerKilled` GameMessage type and handling.
+* Added `NumDeaths`, `DeathLevel`, and `VitaeCpPool` to character properties.
+* Added `PurgeAllEnchantments` GameEvent.
+* Updated `Player.Kill()` to update properties.
 
 ### 2017-04-14
 [Lidefeath]